+++ src/Sqladmin/Operations.ts
@@ -52,10 +52,14 @@ export const makeAwaitOperation = (getOperations: GetOperations) =>
       ),
       Effect.retry({
         while: (e: { _tag?: string }) => e?._tag === "OperationPending",
-        schedule: Schedule.exponential(Duration.seconds(1), 1.5).pipe(
-          Schedule.either(Schedule.spaced(Duration.seconds(10))),
-          Schedule.both(Schedule.recurs(60)),
-          Schedule.tapOutput(() =>
+        schedule: Schedule.max([
+          Schedule.min([
+            Schedule.exponential(Duration.seconds(1), 1.5),
+            Schedule.spaced(Duration.seconds(10)),
+          ]),
+          Schedule.recurs(60),
+        ]).pipe(
+          Schedule.tap(() =>
             session.note(`Waiting for Cloud SQL operation ${operationName}…`),
           ),
         ),